Simplify check_os() code
I've just learned that OBS already embeds LuaJIT, so I don't have to worry about
checking for environment variables and such.  A lot of that code was removed as
a result.

Also, we didn't really need to test for what architecture the system is running
since we're only worried about what OS the user is running.  Architecture
testing code has been removed as well.